chore: #657 move hooks to orders lib
