+++ libs/orders/src/lib/order-hooks/deal-ticket-query.ts
@@ -1,0 +1,33 @@
+import { gql } from '@apollo/client';
+
+export const DEAL_TICKET_QUERY = gql`
+  query DealTicketQuery($marketId: ID!) {
+    market(id: $marketId) {
+      id
+      name
+      decimalPlaces
+      positionDecimalPlaces
+      state
+      tradingMode
+      tradableInstrument {
+        instrument {
+          product {
+            ... on Future {
+              quoteName
+              settlementAsset {
+                id
+                symbol
+                name
+              }
+            }
+          }
+        }
+      }
+      depth {
+        lastTrade {
+          price
+        }
+      }
+    }
+  }
+`;
