+++ libs/orders/src/lib/order-hooks/index.ts
@@ -1,3 +1,5 @@
 export * from './__generated__';
+export * from './deal-ticket-query';
 export * from './order-event-query';
 export * from './use-order-cancel';
+export * from './use-order-edit';
