+++ libs/orders/src/lib/order-hooks/use-order-edit.tsx
@@ -1,18 +1,17 @@
-import { useCallback, useState } from 'react';
-import { determineId } from '@vegaprotocol/react-helpers';
-
-import { useVegaTransaction } from '../use-vega-transaction';
-import { useVegaWallet } from '../use-vega-wallet';
 import { useSubscription } from '@apollo/client';
+import { determineId } from '@vegaprotocol/react-helpers';
+import type { OrderAmendmentBody } from '@vegaprotocol/vegawallet-service-api-client';
+import { useState, useCallback } from 'react';
+import { useVegaTransaction, useVegaWallet } from '@vegaprotocol/wallet';
+import { ORDER_EVENT_SUB } from './order-event-query';
 import type {
+  OrderEvent_busEvents_event_Order,
   OrderEvent,
   OrderEventVariables,
-  OrderEvent_busEvents_event_Order,
-} from './__generated__/OrderEvent';
-import { ORDER_EVENT_SUB } from './order-event-query';
+} from './__generated__';
 import * as Sentry from '@sentry/react';
 
-export const useOrderCancel = () => {
+export const useOrderEdit = () => {
   const { keypair } = useVegaWallet();
   const { send, transaction, reset: resetTransaction } = useVegaTransaction();
   const [updatedOrder, setUpdatedOrder] =
@@ -49,11 +48,15 @@ export const useOrderCancel = () => {
         const res = await send({
           pubKey: keypair.pub,
           propagate: true,
-          orderCancellation: {
+          orderAmendment: {
             orderId: order.id,
             marketId: order.market.id,
+            timeInForce: order.timeInForce,
+            sizeDelta: order.size,
+            price: order.price,
+            expiresAt: order.expiresAt,
           },
-        });
+        } as OrderAmendmentBody);
 
         if (res?.signature) {
           setId(determineId(res.signature));
