+++ libs/wallet/src/manage-dialog/index.ts
@@ -1,0 +1,1 @@
+export * from './manage-dialog';
